new component for datasets that algorithm can run compute job on
