+++ src/components/organisms/AssetActions/Compute/index.module.css
@@ -19,27 +19,3 @@
 .feedback:empty {
   margin-top: 0;
 }
-
-.datasetsContainer {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  align-items: center;
-  width: auto;
-  margin-left: -2rem;
-  margin-right: -2rem;
-}
-
-.datasetsContainer div[class*='AssetSelection-module--selection'] {
-  width: 100%;
-  border-top-left-radius: 0;
-  border-top-right-radius: 0;
-  border-left: 0;
-  border-right: 0;
-  padding: 0;
-}
-
-.datasetsContainer .text {
-  margin-bottom: calc(var(--spacer) / 2);
-  text-align: center;
-}
