+++ src/components/organisms/AssetActions/Consume.tsx
@@ -16,14 +16,7 @@ import { useWeb3 } from '../../../providers/Web3'
 import { usePricing } from '../../../hooks/usePricing'
 import { useConsume } from '../../../hooks/useConsume'
 import ButtonBuy from '../../atoms/ButtonBuy'
-import axios from 'axios'
-import AssetSelection, {
-  AssetSelectionAsset
-} from '../../molecules/FormFields/AssetSelection'
-import {
-  queryMetadata,
-  transformDDOToAssetSelection
-} from '../../../utils/aquarius'
+import AlgorithmDatasetsListForCompute from '../AssetContent/AlgorithmDatasetsListForCompute'
 
 const previousOrderQuery = gql`
   query PreviousOrder($id: String!, $account: String!) {
@@ -51,7 +44,7 @@ export default function Consume({
   dtBalance: string
 }): ReactElement {
   const { accountId } = useWeb3()
-  const { ocean, config } = useOcean()
+  const { ocean } = useOcean()
   const { marketFeeAddress } = useSiteMetadata()
   const [hasPreviousOrder, setHasPreviousOrder] = useState(false)
   const [previousOrderId, setPreviousOrderId] = useState<string>()
@@ -67,48 +60,6 @@ export default function Consume({
   const [hasDatatoken, setHasDatatoken] = useState(false)
   const [isConsumable, setIsConsumable] = useState(true)
   const [assetTimeout, setAssetTimeout] = useState('')
-  const [datasetsForCompute, setDatasetsForCompute] = useState<
-    AssetSelectionAsset[]
-  >()
-  const query = {
-    query: {
-      query_string: {
-        query: `service.attributes.main.privacy.publisherTrustedAlgorithms.did:${ddo.id}`
-      }
-    },
-    sort: { created: -1 }
-  }
-
-  useEffect(() => {
-    async function getDatasetsAllowedForCompute() {
-      const source = axios.CancelToken.source()
-      const computeDatasets = await queryMetadata(
-        query,
-        config.metadataCacheUri,
-        source.token
-      )
-      const computeDatasetsForCurrentAlgorithm: DDO[] = []
-      computeDatasets.results.forEach((data: DDO) => {
-        const algorithm = data
-          .findServiceByType('compute')
-          .attributes.main.privacy.publisherTrustedAlgorithms.find(
-            (algo) => algo.did === ddo.id
-          )
-        algorithm && computeDatasetsForCurrentAlgorithm.push(data)
-      })
-      if (computeDatasetsForCurrentAlgorithm.length === 0) {
-        setDatasetsForCompute([])
-        return
-      }
-      const datasets = await transformDDOToAssetSelection(
-        computeDatasetsForCurrentAlgorithm,
-        config.metadataCacheUri,
-        []
-      )
-      setDatasetsForCompute(datasets)
-    }
-    type === 'algorithm' && getDatasetsAllowedForCompute()
-  }, [type])
 
   const { data } = useQuery<OrdersData>(previousOrderQuery, {
     variables: {
@@ -221,12 +172,7 @@ export default function Consume({
         </div>
       </div>
       {type === 'algorithm' && (
-        <div className={styles.datasetsContainer}>
-          <span className={styles.text}>
-            Datasets algorithm is allowed to run on
-          </span>
-          <AssetSelection assets={datasetsForCompute} hideRadio />
-        </div>
+        <AlgorithmDatasetsListForCompute algorithmDid={ddo.id} />
       )}
       <footer className={styles.feedback}>
         <Web3Feedback isBalanceSufficient={isBalanceSufficient} />
