+++ src/utils/aquarius.ts
@@ -13,6 +13,17 @@ import { AssetSelectionAsset } from '../components/molecules/FormFields/AssetSel
 import { PriceList, getAssetPrices } from './subgraph'
 import axios, { CancelToken, AxiosResponse } from 'axios'
 
+function getQueryForAlgorithmDatasets(algorithmDid: string) {
+  return {
+    query: {
+      query_string: {
+        query: `service.attributes.main.privacy.publisherTrustedAlgorithms.did:${algorithmDid}`
+      }
+    },
+    sort: { created: -1 }
+  }
+}
+
 // TODO: import directly from ocean.js somehow.
 // Transforming Aquarius' direct response is needed for getting actual DDOs
 // and not just strings of DDOs. For now, taken from
@@ -147,4 +158,34 @@ export async function transformDDOToAssetSelection(
     }
   })
   return algorithmList
+}
+
+export async function getAlgorithmDatasetsForCompute(
+  algorithmId: string,
+  metadataCacheUri: string
+): Promise<AssetSelectionAsset[]> {
+  const source = axios.CancelToken.source()
+  const computeDatasets = await queryMetadata(
+    getQueryForAlgorithmDatasets(algorithmId),
+    metadataCacheUri,
+    source.token
+  )
+  const computeDatasetsForCurrentAlgorithm: DDO[] = []
+  computeDatasets.results.forEach((data: DDO) => {
+    const algorithm = data
+      .findServiceByType('compute')
+      .attributes.main.privacy.publisherTrustedAlgorithms.find(
+        (algo) => algo.did === algorithmId
+      )
+    algorithm && computeDatasetsForCurrentAlgorithm.push(data)
+  })
+  if (computeDatasetsForCurrentAlgorithm.length === 0) {
+    return []
+  }
+  const datasets = await transformDDOToAssetSelection(
+    computeDatasetsForCurrentAlgorithm,
+    metadataCacheUri,
+    []
+  )
+  return datasets
 }
